+++ src/main/java/ru/yandex/practicum/filmorate/model/Friends.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @Builder
 public class Friends {
-    private long userId;
-    private long friendId;
+    private int userId;
+    private int friendId;
 }
